+++ LUTCalc/CHANGELOG.md
@@ -1,8 +1,27 @@
 LUTCalc Changelog
 =================
 
+v2.53
+-----
+* New Feature - REDWideGamutRGB and Log3G10.
+
+v2.52
+-----
+* Bugfix - Fixed Canon C-Log3 bug.
+
+v2.51
+-----
+* New Feature - Canon C-Log3.
+
+v2.5
+----
+* New Feature - Black Gamma.
+* New Feature - Sony FS5 camera option.
+* New Feature - Panasonic GH4 V-Log camera option.
+* Bugfix - When using LUTAnalyst, the preview window was not automatically updating to reflect the analysed colour gamut.
+
 v2.4
-------
+----
 * New Feature - LUTAnalyst automatically sets output gamma and gamut to the analysed LUT.
 * New Feature - Alexa Classic 709 (Alexa-X-2) gamma and gamut.
 * New Feature - Amira Rec709 gamma.